+++ VendorOrderTracker/Models/Orders.cs
@@ -6,12 +6,12 @@ namespace VendorOrderTracker.Models
   {
     public string Title { get; set; }
     public string Description { get; set; }
-    public int Price { get; set; }
+    public float Price { get; set; }
     public DateTime Date { get; set; }
     public int Id { get; }
     private static List<Orders> _instances = new List<Orders> { };
 
-    public Orders(string orderTitle, string orderDescription, int orderPrice, DateTime orderDate )
+    public Orders(string orderTitle, string orderDescription, float orderPrice, DateTime orderDate )
     {
       Title = orderTitle;
       Description = orderDescription;
